FIXED GPF on execution of user thread.

+++ kernel/include/arch/x86_64/context.h
@@ -1,51 +1,52 @@
 #pragma once
 
 #include <lib/stdint.h>
+#include <lib/types.h>
 
 typedef struct {
-    uintptr_t fs;
-    uintptr_t ds;
+    u64 fs;
+    u64 ds;
 
-    uintptr_t r15;
-    uintptr_t r14;
-    uintptr_t r13;
-    uintptr_t r12;
-    uintptr_t r11;
-    uintptr_t r10;
-    uintptr_t r9;
-    uintptr_t r8;
+    u64 r15;
+    u64 r14;
+    u64 r13;
+    u64 r12;
+    u64 r11;
+    u64 r10;
+    u64 r9;
+    u64 r8;
 
-    uintptr_t rbp;
-    uintptr_t rsi;
-    uintptr_t rdi;
-    uintptr_t rdx;
-    uintptr_t rcx;
-    uintptr_t rbx;
-    uintptr_t rax;
+    u64 rbp;
+    u64 rsi;
+    u64 rdi;
+    u64 rdx;
+    u64 rcx;
+    u64 rbx;
+    u64 rax;
 
-    uintptr_t trapno;
-    uintptr_t err_code;
+    u64 trapno;
+    u64 err_code;
     
-    uintptr_t rip;
-    uintptr_t cs;
-    uintptr_t rflags;
-    uintptr_t rsp;
-    uintptr_t ss;
+    u64 rip;
+    u64 cs;
+    u64 rflags;
+    u64 rsp;
+    u64 ss;
 }tf_t;
 
 #define x86_64_tf_isuser(tf)    ({((tf)->err_code & 0x4) ? 1 : 0; })
 
 typedef struct {
-    uintptr_t r15;
-    uintptr_t r14;
-    uintptr_t r13;
-    uintptr_t r12;
-    uintptr_t r11;
-    uintptr_t r10;
-    uintptr_t r9;
-    uintptr_t rbx;
-    uintptr_t rbp;
-    uintptr_t rip;
+    u64 r15;
+    u64 r14;
+    u64 r13;
+    u64 r12;
+    u64 r11;
+    u64 r10;
+    u64 r9;
+    u64 rbx;
+    u64 rbp;
+    u64 rip;
 } context_t;
 
 extern void trapret(void);
